Add research notes and minor changes

+++ egd/game/cards.py
@@ -3,7 +3,8 @@ import numpy as np
 
 NUM_CARD_VALUES = 13
 JOKER = 12  # Jokers at index 12
-AVAILABLE_CARDS = np.array([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 2])
+AVAILABLE_CARDS = np.array(
+    [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 2], dtype=np.int8)
 
 
 def get_cards_array(card_type, num_cards):
